# columnToFieldIndex: Improve performance