--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,15 @@
+module github.com/go-gorp/gorp
+
+go 1.13
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-sql-driver/mysql v1.4.1
+	github.com/lib/pq v1.2.0
+	github.com/mattn/go-sqlite3 v1.11.0
+	github.com/poy/onpar v0.0.0-20190519213022-ee068f8ea4d1
+	github.com/stretchr/testify v1.4.0
+	github.com/ziutek/mymysql v1.5.4
+	google.golang.org/appengine v1.6.5 // indirect
+	gopkg.in/yaml.v2 v2.2.4 // indirect
+)