This PR is in the same spirit as https://github.com/go-gorp/gorp/pull/347

We found that the function `columnToFieldIndex` was inefficient from a CPU usage POV, but more importantly created ton of objects for the GC.

Analyzing real world usage of gorp in a proprietary API server, we were able to reduce overall memory allocs by ~50% and bytes allocated by ~20% with this patch. 

This PR:
- Adds a few tests and a micro-benchmark for `columnToFieldIndex`
- Takes a dependency on `reflectx` to implement `columnToFieldIndex`.  #347 tries to implement it locally, but `reflectx` does exactly what we need here, and has built-in caching based on the `reflect.Type`.  We would need to replicate a few things into gorp if the dependency is not acceptable.


Micro-benchmark results:
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCcolumnToFieldIndex-8     7183          334           -95.35%

benchmark                          old allocs     new allocs     delta
BenchmarkCcolumnToFieldIndex-8     77             2              -97.40%

benchmark                          old bytes     new bytes     delta
BenchmarkCcolumnToFieldIndex-8     1256          208           -83.44%
```