@paulquerna-okta 2019-10-31
(note, I think we need some improved tests and changes for embedded structures, but I hope this is still a directionally interesting PR)