@hi-wayne 2021-01-25
sqlx 1.2 reflectx  not suppor  embedded structs
fyi sqlx1.2 release
https://github.com/jmoiron/sqlx/pull/429